requirements.txt
fastapi
uvicorn
openai
python-telegram-bot
langchain
python-dotenv
requests
